refactor(modular Hamiltonian): unify Hamiltonian class
Lift the splitting of the Hamiltonian class into a `Hamiltonian_Heisenberg`
and a `Hamiltonian_Gaussian` in favor of a unified `Hamiltonian` class.
The `Gaussian` part will be reintroduced as an Interaction class that is
stored inside the new `Hamiltonian` class if needed. The code for this
implementation is already present, but the `Gaussian` functionality is
neutered in this commit.