decouple `utility/Configurations` from the `Spin_System` class
In order to make these operations more generic the functions now
operate on a pair of a `vectorfield` reference and a constant
`Data::Geometry` reference. The solution isn't optimal, because the
dimensionality of the `vectorfield` and `Geometry` have to match, which
was ensured before by these two dependencies belonging to the same
`Spin_System` object.
Maybe adding a transactional class that carries references to both the
spin configuration (`vectorfield`) and the associated `Geometry` data
could be used instead, but this option also works for now.