Add dynamics test for the pinning feature
The pinning feature was previously not captured in our test suite.
This commit adds a test for the dynamics of a spin precessing under
the effective magnetic field of an exchange coupled, pinned spin.
Parameters have been chosen to match the precession frequency of the
Lamor precession test from `test_physics` which uses only the Zeeman
term.